+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The changelog for `Superwall`. Also see the [releases](https://github.com/superwall/Superwall-Android/releases) on GitHub.
 
+## 2.5.7
+
+## Fixes
+- Fix `demandScore` and `demandTier` getting removed from some events
+- Fixes paywall navigation resetting after backgrounding
+- Removes webview flags which can cause off-screen render issues
+
 ## 2.5.6
 
 ## Enhancements